requirements.txt
streamlit
easyocr
opencv-python
numpy
pillow